requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
WTForms==3.0.1
email-validator==2.1.0
requests==2.31.0
python-dotenv==1.0.0
reportlab==4.0.4
APScheduler==3.10.4
Flask-Bootstrap==3.3.7.1
Pillow==10.0.1
Werkzeug==2.3.7
